More clearly differentiate core and api's SQL adapters
