+++ api/dbadapters/index.ts
@@ -1,7 +1,7 @@
 import { Credentials } from "df/api/commands/credentials";
 import { BigQueryDbAdapter } from "df/api/dbadapters/bigquery";
-import { QueryOrAction } from "df/core/adapters";
 import { dataform } from "df/protos/ts";
+import { QueryOrAction } from "df/api/dbadapters/execution_sql_adapter";
 
 export type OnCancel = (handleCancel: () => void) => void;
