+++ tests/integration/utils.ts
@@ -2,7 +2,7 @@ import { expect } from "chai";
 
 import * as dfapi from "df/api";
 import * as dbadapters from "df/api/dbadapters";
-import * as adapters from "df/core/adapters";
+import { ExecutionSqlAdapter } from "df/api/dbadapters/execution_sql_adapter";
 import { dataform } from "df/protos/ts";
 
 export function keyBy<V>(values: V[], keyFn: (value: V) => string): { [key: string]: V } {
@@ -14,20 +14,23 @@ export function keyBy<V>(values: V[], keyFn: (value: V) => string): { [key: stri
 
 export async function dropAllTables(
   tables: dataform.ITableMetadata[],
-  adapter: adapters.IAdapter,
+  executionSqlAdapter: ExecutionSqlAdapter,
   dbadapter: dbadapters.IDbAdapter
 ) {
   await Promise.all(
-    tables.map(table => dbadapter.execute(adapter.dropIfExists(table.target, table.type)))
+    tables.map(table =>
+      dbadapter.execute(executionSqlAdapter.dropIfExists(table.target, table.type))
+    )
   );
 }
 
 export async function getTableRows(
   target: dataform.ITarget,
-  adapter: adapters.IAdapter,
+  executionSqlAdapter: ExecutionSqlAdapter,
   dbadapter: dbadapters.IDbAdapter
 ) {
-  return (await dbadapter.execute(`SELECT * FROM ${adapter.resolveTarget(target)}`)).rows;
+  return (await dbadapter.execute(`SELECT * FROM ${executionSqlAdapter.resolveTarget(target)}`))
+    .rows;
 }
 
 export async function compile(
